requirements.txt
flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0
pandas==1.5.3
numpy==1.24.3
langchain==0.0.350
langchain-google-genai==0.0.5
python-dotenv==1.0.0
plotly==5.18.0
werkzeug==2.0.3
pydantic==1.10.13
openpyxl==3.1.2 